Use float scaling

+++ movie_player.go
@@ -47,7 +47,7 @@ func (p *moviePlayer) initUI() {
 	win.Closable = false
 	win.Resizable = false
 	win.AutoSize = true
-	win.Position = eui.Point{X: float32((gameAreaSizeX * gs.Scale) / 2), Y: 5}
+	win.Position = eui.Point{X: float32(float64(gameAreaSizeX) * gs.Scale / 2), Y: 5}
 	win.Closable = false
 
 	flow := &eui.ItemData{ItemType: eui.ITEM_FLOW, FlowType: eui.FLOW_VERTICAL}
